updates to SoilErosionAlgorithm
git-svn-id: http://tikouka.googlecode.com/svn/trunk/52n-wps-tikouka@9 88633496-3efa-8c45-26c6-2b4ef57d3061

+++ src/main/java/tikouka/nl/wps/algorithm/util/lookuptable.java
@@ -13,7 +13,7 @@ public class lookuptable {
 
     private String id;
     private String key;
-    private double value;
+    private Integer value;
 
     public String getId() {
         return id;
@@ -35,7 +35,7 @@ public class lookuptable {
         this.key = key;
     }
 
-    public void setValue(double value) {
+    public void setValue(Integer value) {
         this.value = value;
     }
 
@@ -44,7 +44,7 @@ public class lookuptable {
         key = "";
     }
 
-    public lookuptable(String pId, String pKey, double pValue) {
+    public lookuptable(String pId, String pKey, Integer pValue) {
         id = pId;
         key = pKey;
         value = pValue;
